Refactor Instagram profile follow-up runs into separate function
