+++ lib/apify/handleInstagramProfileScraperResults.ts
@@ -14,8 +14,7 @@ import getAccountEmails from "../supabase/accountEmails/getAccountEmails";
 import sendApifyWebhookEmail from "@/lib/apify/sendApifyWebhookEmail";
 import normalizeProfileUrl from "@/lib/utils/normalizeProfileUrl";
 import uploadLinkToArweave from "@/lib/arweave/uploadLinkToArweave";
-import runInstagramCommentsScraper from "@/lib/apify/runInstagramCommentsScraper";
-import getExistingPostComments from "@/lib/apify/getExistingPostComments";
+import handleInstagramProfileFollowUpRuns from "@/lib/apify/handleInstagramProfileFollowUpRuns";
 import apifyPayloadSchema from "@/lib/apify/apifyPayloadSchema";
 
 /**
@@ -91,32 +90,7 @@ export default async function handleInstagramProfileScraperResults(
         );
 
         // Trigger comment scraping for the new posts
-        // Only call runInstagramCommentsScraper if dataset.length === 1
-        // If more than 1 profile, these are fans and should only be saved
-        if (dataset.length === 1 && firstResult.latestPosts && firstResult.latestPosts.length > 0) {
-          const postUrls = (firstResult.latestPosts as ApifyInstagramPost[])
-            .map((post) => post.url)
-            .filter(Boolean);
-          
-          if (postUrls.length > 0) {
-            console.log("Checking existing comments for posts:", postUrls);
-            
-            // Get existing comments for these post URLs
-            const { urlsWithComments, urlsWithoutComments } = await getExistingPostComments(postUrls);
-            
-            // Handle posts with existing comments (use resultsLimit)
-            if (urlsWithComments.length > 0) {
-              console.log("Triggering comment scraping for posts with existing comments (resultsLimit=1):", urlsWithComments);
-              await runInstagramCommentsScraper(urlsWithComments, 1);
-            }
-            
-            // Handle posts without existing comments (no resultsLimit)
-            if (urlsWithoutComments.length > 0) {
-              console.log("Triggering comment scraping for posts without existing comments:", urlsWithoutComments);
-              await runInstagramCommentsScraper(urlsWithoutComments);
-            }
-          }
-        }
+        await handleInstagramProfileFollowUpRuns(dataset, firstResult);
       }
     }
   }
